requirements.txt
PyYAML
imagesize
tqdm